wip on finish sorting
