+++ dbms/src/Core/SortDescription.h
@@ -26,6 +26,17 @@ struct SortColumnDescription
 
     SortColumnDescription(const std::string & column_name_, int direction_, int nulls_direction_, const std::shared_ptr<Collator> & collator_ = nullptr)
         : column_name(column_name_), column_number(0), direction(direction_), nulls_direction(nulls_direction_), collator(collator_) {}
+
+    bool operator == (const SortColumnDescription & other) const
+    {
+        return column_name == other.column_name && column_number == other.column_number
+            && direction == other.direction && nulls_direction == other.nulls_direction;
+    }
+
+    bool operator != (const SortColumnDescription & other) const
+    {
+        return !(*this == other);
+    }
 };
 
 /// Description of the sorting rule for several columns.
